+++ idkit/src/components/IDKitWidget/States/WorldIDState.tsx
@@ -1,3 +1,4 @@
+import { IDKITStage } from '@/types'
 import { getCopy } from '@/lib/utils'
 import useMedia from '@/hooks/useMedia'
 import QRState from './WorldID/QRState'
@@ -5,42 +6,44 @@ import useIDKitStore from '@/store/idkit'
 import { useEffect, useState } from 'react'
 import { VerificationState } from '@/types/orb'
 import type { IDKitStore } from '@/store/idkit'
-import { IDKITStage, SignalType } from '@/types'
 import useOrbSignal from '@/services/walletconnect'
 import AboutWorldID from '@/components/AboutWorldID'
 import LoadingIcon from '@/components/Icons/LoadingIcon'
-import DevicePhoneMobileIcon from '@/components/Icons/DevicePhoneMobileIcon'
 
 const getOptions = (store: IDKitStore) => ({
 	signal: store.signal,
 	copy: store.copy,
-	actionId: store.actionId,
+	app_id: store.app_id,
+	action: store.action,
+	action_description: store.action_description,
+	walletConnectProjectId: store.walletConnectProjectId,
 	handleVerify: store.handleVerify,
-	showAbout: store.methods.length == 1,
-	hasPhone: store.methods.includes('phone'),
-	usePhone: () => store.setStage(IDKITStage.ENTER_PHONE),
+	setStage: store.setStage,
 })
 
 const WorldIDState = () => {
 	const media = useMedia()
 	const [showQR, setShowQR] = useState<boolean>(false)
-	const { actionId, copy, signal, handleVerify, showAbout, hasPhone, usePhone } = useIDKitStore(getOptions)
-	const { result, qrData, verificationState, reset } = useOrbSignal(actionId, signal)
+	const { copy, app_id, action, signal, setStage, handleVerify, action_description, walletConnectProjectId } =
+		useIDKitStore(getOptions)
+
+	const { result, qrData, verificationState, reset } = useOrbSignal(
+		app_id,
+		action,
+		signal,
+		action_description,
+		walletConnectProjectId
+	)
 
 	useEffect(() => reset, [reset])
 
 	useEffect(() => {
-		if (!result) return
+		if (verificationState === VerificationState.Failed) {
+			setStage(IDKITStage.ERROR)
+		}
 
-		handleVerify({
-			signal_type: SignalType.Orb,
-			nullifier_hash: result.nullifier_hash,
-			proof_payload: {
-				proof: result.proof,
-				merkle_root: result.merkle_root,
-			},
-		})
-	}, [result, reset, handleVerify])
+		if (result) handleVerify(result)
+	}, [result, reset, handleVerify, verificationState, setStage])
 
 	return (
 		<div className="-mt-6 space-y-6">
@@ -64,22 +67,7 @@ const WorldIDState = () => {
 			) : (
 				<QRState showQR={showQR} setShowQR={setShowQR} qrData={qrData} />
 			)}
-			{showAbout && (media == 'desktop' || !showQR) && <AboutWorldID />}
-			{hasPhone && verificationState == VerificationState.AwaitingConnection && (
-				<div className="hidden space-y-3 md:block">
-					<div className="flex items-center justify-between space-x-6">
-						<div className="bg-f2f5f9 dark:bg-29343f h-px flex-1" />
-						<p className="text-9eafc0 dark:text-596673 text-xs">or</p>
-						<div className="bg-f2f5f9 dark:bg-29343f h-px flex-1" />
-					</div>
-					<div className="flex items-center justify-center">
-						<button onClick={usePhone} className="flex items-center space-x-2">
-							<DevicePhoneMobileIcon className="text-0d151d/70 h-6 w-6 dark:text-white/70" />
-							<p className="text-0d151d text-sm font-medium dark:text-white">Verify with Phone Number</p>
-						</button>
-					</div>
-				</div>
-			)}
+			{(media == 'desktop' || !showQR) && <AboutWorldID />}
 		</div>
 	)
 }