+++ idkit/src/index.ts
@@ -1,7 +1,16 @@
 import useIDKit from './hooks/useIDKit'
+import type { ISuccessResult } from '@/types'
+import { solidityEncode } from './lib/hashing'
 import IDKitWidget from '@/components/IDKitWidget'
 import type { WidgetProps, Config } from '@/types/config'
-import type { ISuccessResult, SignalType, PhoneSignalProof, OrbSignalProof } from '@/types'
+import { hashToField, validateABILikeEncoding, generateExternalNullifier } from './lib/hashing'
 
-export { IDKitWidget, useIDKit }
-export type { ISuccessResult, Config, SignalType, PhoneSignalProof, OrbSignalProof, WidgetProps }
+// NOTE: For internal use in other Worldcoin packages
+const internal = {
+	hashToField,
+	validateABILikeEncoding,
+	generateExternalNullifier,
+}
+
+export { IDKitWidget, useIDKit, solidityEncode, internal }
+export type { ISuccessResult, Config, WidgetProps }