+++ idkit/src/services/walletconnect.ts
@@ -2,39 +2,47 @@ import { create } from 'zustand'
 import { useEffect } from 'react'
 import { buildQRData } from '@/lib/qr'
 import { randomNumber } from '@/lib/utils'
-import type { OrbResponse } from '@/types/orb'
-import WalletConnect from '@walletconnect/client'
-import type { ExpectedErrorResponse } from '@/types'
-import type { StringOrAdvanced } from '@/types/config'
+import { WC_PROJECT_ID } from '@/lib/consts'
+import Client from '@walletconnect/sign-client'
+import type { IDKitConfig } from '@/types/config'
+import { getSdkError } from '@walletconnect/utils'
 import { OrbErrorCodes, VerificationState } from '@/types/orb'
-import { validateABILikeEncoding, worldIDHash } from '@/lib/hashing'
+import type { ExpectedErrorResponse, ISuccessResult } from '@/types'
+import { validateABILikeEncoding, generateSignal, generateExternalNullifier, encodeAction } from '@/lib/hashing'
 
 type WalletConnectStore = {
 	connected: boolean
-	connectorUri: string
-	result: OrbResponse | null
+	uri: string
+	topic: string
+	result: ISuccessResult | null
 	errorCode: OrbErrorCodes | null
 	verificationState: VerificationState
-	config: { action_id: StringOrAdvanced; signal: StringOrAdvanced } | null
+	config: IDKitConfig | null
 	qrData: {
 		default: string
 		mobile: string
 	} | null
+	client: Client | null
 
 	resetConnection: () => void
-	onConnectionEstablished: () => void
-	setConnectorUri: (uri: string) => void
-	initConnection: (action_id: StringOrAdvanced, signal: StringOrAdvanced) => Promise<void>
+	onConnectionEstablished: (client: Client) => void
+	setUri: (uri: string) => void
+	createClient: (
+		app_id: IDKitConfig['app_id'],
+		action: IDKitConfig['action'],
+		signal: IDKitConfig['signal'],
+		action_description?: IDKitConfig['action_description'],
+		walletConnectProjectId?: IDKitConfig['walletConnectProjectId']
+	) => Promise<void>
+	connectClient: (client: Client) => Promise<void>
 }
 
-let connector: WalletConnect
-
-try {
-	connector = new WalletConnect({
-		bridge: 'https://bridge.walletconnect.org',
-	})
-} catch (error) {
-	console.error('Unable to create WalletConnect connector')
+const namespaces = {
+	eip155: {
+		methods: ['world_id_v1'],
+		chains: ['eip155:1'], // Chain ID used does not matter, since we only perform custom JSON RPC messages (World ID verification)
+		events: ['accountsChanged'],
+	},
 }
 
 const useWalletConnectStore = create<WalletConnectStore>()((set, get) => ({
@@ -42,52 +50,88 @@ const useWalletConnectStore = create<WalletConnectStore>()((set, get) => ({
 	config: null,
 	result: null,
 	connected: false,
-	connectorUri: '',
+	uri: '',
+	topic: '',
 	errorCode: null,
 	verificationState: VerificationState.LoadingWidget,
+	client: null,
 
-	initConnection: async (action_id: StringOrAdvanced, signal: StringOrAdvanced) => {
-		if (connector.connected) await connector.killSession()
+	createClient: async (
+		app_id: IDKitConfig['app_id'],
+		action: IDKitConfig['action'],
+		signal: IDKitConfig['signal'],
+		action_description?: IDKitConfig['action_description'],
+		walletConnectProjectId = WC_PROJECT_ID
+	) => {
+		set({ config: { app_id, action, signal, action_description, walletConnectProjectId } })
 
-		set({ config: { action_id, signal } })
+		try {
+			const client = await Client.init({
+				projectId: walletConnectProjectId,
+				metadata: {
+					name: 'IDKit',
+					description: 'Verify with World ID',
+					url: '#',
+					icons: ['https://worldcoin.org/icons/logo-small.svg'],
+				},
+			})
+			set({ client: client })
 
-		await connector.createSession()
-		get().setConnectorUri(connector.uri)
-
-		connector.on('connect', (error: unknown) => {
-			if (!error) return get().onConnectionEstablished()
-
-			set({ errorCode: OrbErrorCodes.ConnectionFailed })
-			console.error(`Unable to establish a connection with the WLD app: ${error}`)
-		})
-
-		connector.on('disconnect', () => {
-			void get().initConnection(action_id, signal)
-		})
+			return get().connectClient(client)
+		} catch (error) {
+			console.error(error)
+		}
 	},
 
-	setConnectorUri: (uri: string) => {
+	connectClient: async (client: Client) => {
+		try {
+			const { uri, approval } = await client.connect({ requiredNamespaces: namespaces })
+
+			if (uri) {
+				get().setUri(uri)
+				const session = await approval()
+
+				if (typeof session !== 'undefined') {
+					set({ topic: session.topic })
+					return get().onConnectionEstablished(client)
+				}
+			}
+		} catch (error) {
+			set({ errorCode: OrbErrorCodes.ConnectionFailed })
+			console.error(`Unable to establish a connection with the WLD app: ${error}`)
+		}
+	},
+
+	setUri: (uri: string) => {
 		if (!uri) return
 
 		set({
-			connectorUri: uri,
+			uri: uri,
 			verificationState: VerificationState.AwaitingConnection,
 			qrData: {
-				default: buildQRData(connector),
-				mobile: buildQRData(connector, window.location.href),
+				default: buildQRData(uri),
+				mobile: buildQRData(uri, window.location.href),
 			},
 		})
 	},
-	onConnectionEstablished: () => {
+
+	// eslint-disable-next-line @typescript-eslint/no-misused-promises
+	onConnectionEstablished: async (client: Client) => {
 		set({ verificationState: VerificationState.AwaitingVerification })
 
-		connector
-			// eslint-disable-next-line @typescript-eslint/no-non-null-assertion
-			.sendCustomRequest(buildVerificationRequest(get().config!.action_id, get().config!.signal))
-			.then((result: Record<string, string | undefined>) => {
-				if (!ensureVerificationResponse(result)) return set({ errorCode: OrbErrorCodes.UnexpectedResponse })
+		await client
+			.request({
+				topic: get().topic,
+				chainId: 'eip155:1',
+				// eslint-disable-next-line @typescript-eslint/no-non-null-assertion
+				request: buildVerificationRequest(get().config!),
+			})
+			.then(result => {
+				if (!ensureVerificationResponse(result)) {
+					return set({ errorCode: OrbErrorCodes.UnexpectedResponse })
+				}
 
-				set({ result, verificationState: VerificationState.Confirmed })
+				set({ result: result, verificationState: VerificationState.Confirmed })
 			})
 			.catch((error: unknown) => {
 				let errorCode = OrbErrorCodes.GenericError
@@ -99,8 +143,13 @@ const useWalletConnectStore = create<WalletConnectStore>()((set, get) => ({
 
 				set({ errorCode, verificationState: VerificationState.Failed })
 			})
-			.finally(() => void connector.killSession())
-			.catch(error => console.error('Unable to kill session', error))
+			// eslint-disable-next-line @typescript-eslint/no-misused-promises
+			.finally(async () => {
+				await client.disconnect({ topic: get().topic, reason: getSdkError('USER_DISCONNECTED') })
+			})
+			.catch(error => {
+				console.error(`Unable to disconnect: ${error}`)
+			})
 	},
 
 	resetConnection: () => {
@@ -110,24 +159,35 @@ const useWalletConnectStore = create<WalletConnectStore>()((set, get) => ({
 			qrData: null,
 			errorCode: null,
 			connected: false,
-			connectorUri: '',
+			client: null,
 			verificationState: VerificationState.LoadingWidget,
 		})
 	},
 }))
 
-const buildVerificationRequest = (action_id: StringOrAdvanced, signal: StringOrAdvanced) => ({
+const buildVerificationRequest = (config: IDKitConfig) => ({
 	jsonrpc: '2.0',
-	method: 'wld_worldIDVerification',
+	method: 'world_id_v1',
 	id: randomNumber(100000, 9999999),
-	params: [{ signal: worldIDHash(signal).digest, action_id: worldIDHash(action_id).digest }],
+	params: [
+		{
+			app_id: config.app_id,
+			action: encodeAction(config.action),
+			signal: generateSignal(config.signal).digest,
+			action_description: config.action_description,
+			external_nullifier: generateExternalNullifier(config.app_id, config.action).digest,
+		},
+	],
 })
 
-const ensureVerificationResponse = (result: Record<string, string | undefined>): result is OrbResponse => {
-	const proof = 'proof' in result ? result.proof : undefined
-	const merkle_root = 'merkle_root' in result ? result.merkle_root : undefined
-	const nullifier_hash = 'nullifier_hash' in result ? result.nullifier_hash : undefined
+const ensureVerificationResponse = (result: unknown): result is ISuccessResult => {
+	if (!result || typeof result !== 'object') return false
+	const proof = 'proof' in result ? (result as Record<string, string>).proof : undefined
+	const merkle_root = 'merkle_root' in result ? (result as Record<string, string>).merkle_root : undefined
+	const nullifier_hash = 'nullifier_hash' in result ? (result as Record<string, string>).nullifier_hash : undefined
+	const credential_type = 'credential_type' in result ? (result as Record<string, string>).credential_type : undefined
 
+	if (!credential_type) return false
 	for (const attr of [merkle_root, nullifier_hash, proof]) {
 		if (!attr || !validateABILikeEncoding(attr)) return false
 	}
@@ -137,7 +197,7 @@ const ensureVerificationResponse = (result: Record<string, string | undefined>):
 
 type UseOrbSignalResponse = {
 	reset: () => void
-	result: OrbResponse | null
+	result: ISuccessResult | null
 	errorCode: OrbErrorCodes | null
 	verificationState: VerificationState
 	qrData: {
@@ -151,18 +211,28 @@ const getStore = (store: WalletConnectStore) => ({
 	result: store.result,
 	errorCode: store.errorCode,
 	reset: store.resetConnection,
-	initConnection: store.initConnection,
+	client: store.client,
+	createClient: store.createClient,
+	connectClient: store.connectClient,
 	verificationState: store.verificationState,
 })
 
-const useOrbSignal = (action_id: StringOrAdvanced, signal: StringOrAdvanced): UseOrbSignalResponse => {
-	const { result, verificationState, errorCode, qrData, initConnection, reset } = useWalletConnectStore(getStore)
+const useOrbSignal = (
+	app_id: IDKitConfig['app_id'],
+	action: IDKitConfig['action'],
+	signal: IDKitConfig['signal'],
+	action_description?: IDKitConfig['action_description'],
+	walletConnectProjectId?: IDKitConfig['walletConnectProjectId']
+): UseOrbSignalResponse => {
+	const { result, verificationState, errorCode, qrData, client, createClient, reset } =
+		useWalletConnectStore(getStore)
 
 	useEffect(() => {
-		if (!action_id || !signal) return
-
-		void initConnection(action_id, signal)
-	}, [action_id, initConnection, signal])
+		if (!app_id || !action || !signal) return
+		if (!client) {
+			void createClient(app_id, action, signal, action_description, walletConnectProjectId)
+		}
+	}, [app_id, action, signal, walletConnectProjectId, action_description, client, createClient, verificationState])
 
 	return { result, reset, verificationState, errorCode, qrData }
 }