+++ idkit/src/store/idkit.ts
@@ -1,11 +1,16 @@
 import { create } from 'zustand'
-import { worldIDHash } from '@/lib/hashing'
 import { ErrorCodes, IDKITStage } from '@/types'
 import { telemetryModalOpened } from '@/lib/telemetry'
-import type { CallbackFn, ISuccessResult, IErrorState } from '@/types'
-import type { Config, ConfigSource, StringOrAdvanced, VerificationMethods } from '@/types/config'
+import type { CallbackFn, IErrorState, ISuccessResult } from '@/types'
+import type { Config, ConfigSource, IDKitConfig } from '@/types/config'
 
 export type IDKitStore = {
+	app_id: IDKitConfig['app_id']
+	action: IDKitConfig['action']
+	signal: IDKitConfig['signal']
+	action_description?: IDKitConfig['action_description']
+	walletConnectProjectId?: IDKitConfig['walletConnectProjectId']
+
 	code: string
 	open: boolean
 	stage: IDKITStage
@@ -14,18 +19,13 @@ export type IDKitStore = {
 	processing: boolean
 	copy: Config['copy']
 	theme: Config['theme']
-	signal: StringOrAdvanced
-	actionId: StringOrAdvanced
-	stringifiedActionId: string // Raw action IDs get hashed and stored (used for phone non-orb signals)
 	result: ISuccessResult | null
-	methods: VerificationMethods[]
 	errorState: IErrorState | null
 	verifyCallbacks: Record<ConfigSource, CallbackFn | undefined> | Record<string, never>
 	successCallbacks: Record<ConfigSource, CallbackFn | undefined> | Record<string, never>
 
 	computed: {
 		canGoBack: (stage: IDKITStage) => boolean
-		getDefaultStage: (methods?: Config['methods']) => IDKITStage
 	}
 
 	retryFlow: () => void
@@ -42,12 +42,18 @@ export type IDKitStore = {
 }
 
 const useIDKitStore = create<IDKitStore>()((set, get) => ({
+	app_id: '',
+	signal: '',
+	action: '',
+	action_description: '',
+	walletConnectProjectId: '',
+
 	open: false,
 	code: '',
-	signal: '',
 	result: null,
-	actionId: '',
 	theme: 'light',
+	errorTitle: '',
+	errorDetail: '',
 	phoneNumber: '',
 	autoClose: false,
 	errorState: null,
@@ -56,28 +62,17 @@ const useIDKitStore = create<IDKitStore>()((set, get) => ({
 	successCallbacks: {},
 	stringifiedActionId: '',
 	methods: ['orb', 'phone'],
-	stage: IDKITStage.SELECT_METHOD,
+	stage: IDKITStage.WORLD_ID,
 	copy: {},
 
 	computed: {
 		canGoBack: (stage: IDKITStage) => {
 			switch (stage) {
-				case IDKITStage.ENTER_PHONE:
-					return get().methods.includes('orb')
-				case IDKITStage.WORLD_ID:
-					return get().methods.includes('phone')
-				case IDKITStage.ENTER_CODE:
 				case IDKITStage.PRIVACY:
 					return true
 				default:
 					return false
 			}
-		},
-		getDefaultStage: (methods?: Config['methods']) => {
-			methods = methods ?? get().methods
-
-			if (methods.length > 1) return IDKITStage.SELECT_METHOD
-			return methods[0] === 'phone' ? IDKITStage.ENTER_PHONE : IDKITStage.WORLD_ID
 		},
 	},
 
@@ -86,7 +81,9 @@ const useIDKitStore = create<IDKitStore>()((set, get) => ({
 	setErrorState: errorState => set({ errorState }),
 	setPhoneNumber: phoneNumber => set({ phoneNumber }),
 	setProcessing: (processing: boolean) => set({ processing }),
-	retryFlow: () => set({ stage: IDKITStage.ENTER_PHONE, phoneNumber: '', errorState: null }),
+	retryFlow: () => {
+		set({ stage: IDKITStage.WORLD_ID, errorState: null })
+	},
 	addSuccessCallback: (cb: CallbackFn, source: ConfigSource) => {
 		set(state => {
 			state.successCallbacks[source] = cb
@@ -102,18 +99,28 @@ const useIDKitStore = create<IDKitStore>()((set, get) => ({
 		})
 	},
 	setOptions: (
-		{ handleVerify, onSuccess, signal, actionId, autoClose, copy, theme, methods }: Config,
+		{
+			handleVerify,
+			onSuccess,
+			signal,
+			action,
+			app_id,
+			action_description,
+			walletConnectProjectId,
+			autoClose,
+			copy,
+			theme,
+		}: Config,
 		source: ConfigSource
 	) => {
-		const stringifiedActionId = typeof actionId === 'string' ? actionId : worldIDHash(actionId).digest
 		set(store => ({
 			theme,
 			signal,
-			actionId,
-			methods: methods ?? store.methods,
-			stage: store.computed.getDefaultStage(methods),
+			action,
+			app_id,
 			autoClose,
-			stringifiedActionId,
+			action_description,
+			walletConnectProjectId,
 			copy: { ...store.copy, ...copy },
 		}))
 
@@ -160,7 +167,7 @@ const useIDKitStore = create<IDKitStore>()((set, get) => ({
 			phoneNumber: '',
 			errorState: null,
 			processing: false,
-			stage: get().computed.getDefaultStage(),
+			stage: IDKITStage.WORLD_ID,
 		})
 	},
 }))