+++ idkit/src/types/config.ts
@@ -1,22 +1,25 @@
-import type { CallbackFn } from '.'
+import type { CallbackFn, AbiEncodedValue } from '.'
 
-export type VerificationMethods = 'orb' | 'phone'
-export type StringOrAdvanced = Array<[string, unknown]> | string
 export enum ConfigSource {
 	HOOK = 'hook',
 	PROPS = 'props',
 	MANUAL = 'manual',
 }
 
-export type Config = {
+export type IDKitConfig = {
+	app_id: string
+	action_description?: string
+	walletConnectProjectId?: string
+	signal: AbiEncodedValue | string
+	action: AbiEncodedValue | string
+}
+
+type WidgetConfig = {
 	autoClose?: boolean
 	onSuccess?: CallbackFn
 	theme?: 'dark' | 'light'
-	signal: StringOrAdvanced
 	enableTelemetry?: boolean
 	handleVerify?: CallbackFn
-	actionId: StringOrAdvanced
-	methods?: VerificationMethods[]
 	copy?: {
 		title?: string
 		heading?: string
@@ -24,6 +27,8 @@ export type Config = {
 		success?: string
 	}
 }
+
+export type Config = IDKitConfig & WidgetConfig
 
 export type WidgetProps = Config & {
 	children?: ({ open }: { open: () => void }) => JSX.Element